Add files via upload
I wrote the code and when I finished I checked the output but it came up with a white screen and in the top right corner there was a sign that said loading. I checked the console and it said that there was a problem with line 3 and 1 with p5.min.js even though I haven't even used it. I checked Visual Studio and there isn't even a file called p5.min.js.